rename files to follow class name
