+++ src/gui/desktop/quick_views/CMakeLists.txt
@@ -10,8 +10,8 @@ qt6_add_resources(QML_RESOURCES
 )
 
 qt6_wrap_cpp(MOCED
+                amedia_item.hpp
                 bridge.hpp
-                media_item.hpp
                 photo_data_qml.hpp
                 photo_item.hpp
                 photos_model_controller_component.hpp
@@ -23,10 +23,10 @@ qt6_wrap_cpp(MOCED
 add_library(quick_views OBJECT
                 ${QML_RESOURCES}
                 ${MOCED}
+                amedia_item.cpp
+                amedia_item.hpp
                 bridge.cpp
                 bridge.hpp
-                media_item.cpp
-                media_item.hpp
                 photo_data_qml.cpp
                 photo_data_qml.hpp
                 photo_item.cpp
