+++ src/gui/desktop/quick_views/amedia_item.cpp
@@ -1,5 +1,5 @@
 
-#include "media_item.hpp"
+#include "amedia_item.hpp"
 
 
 AMediaItem::AMediaItem(QQuickItem* p)
