+++ src/gui/desktop/quick_views/amedia_item.hpp
